+++ source/extensions/omni.isaac.orbit_tasks/omni/isaac/orbit_tasks/classic/humanoid/__init__.py
@@ -7,21 +7,23 @@
 Humanoid locomotion environment (similar to OpenAI Gym Humanoid-v2).
 """
 
-# import gymnasium as gym
+import gymnasium as gym
 
-# from . import agents
+from . import agents, humanoid_env_cfg
 
 ##
 # Register Gym environments.
 ##
 
-# gym.register(
-#     id="Isaac-Humanoid-v0",
-#     entry_point="omni.isaac.orbit.envs:RLTaskEnv",
-#     kwargs={
-#         "env_cfg_entry_point": f"{__name__}:ant_env_cfg.yaml",
-#         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
-#         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
-#         "sb3_cfg_entry_point": f"{agents.__name__}:sb3_ppo_cfg.yaml",
-#     },
-# )
+gym.register(
+    id="Isaac-Humanoid-v0",
+    entry_point="omni.isaac.orbit.envs:RLTaskEnv",
+    disable_env_checker=True,
+    kwargs={
+        "env_cfg_entry_point": humanoid_env_cfg.HumanoidEnvCfg,
+        "rsl_rl_cfg_entry_point": agents.rsl_rl_ppo_cfg.HumanoidPPORunnerCfg,
+        "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
+        "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
+        "sb3_cfg_entry_point": f"{agents.__name__}:sb3_ppo_cfg.yaml",
+    },
+)